Refactor Settings editor colours into own registry
